Confirmation dialog

+++ src/app/services/basket.service.ts
@@ -7,6 +7,7 @@ import { Utils } from "../helpers/utils";
 import { OrderService } from "./products/order.service";
 import { FoodOrderService } from "./food-order.service";
 import { ToastService } from "./toast.service";
+import { ConfirmationDialogService } from "../shared/confirmation-dialog/confirmation.dialog.service";
 
 @Injectable({
     providedIn: 'root',
@@ -17,28 +18,37 @@ export class BasketService {
     fOrderService = inject(FoodOrderService);
     localService = inject(LocalService);
     toastService = inject(ToastService);
+    confirmationDialogService = inject(ConfirmationDialogService);
     utils = inject(Utils);
 
-    addToProductOrder(OrderItem: OrderItem){
+    addToProductOrder(OrderItem: OrderItem) {
         console.log('Adding a product to cart');
         var json = this.localService.getData(Constants.StorageItem_F_Order);
-        if ( Utils.isValid(json) && Utils.isJsonString(json)){
-            this.toastService.warning("You have a active food cart. Please be informed that food cart will be removed.");
-            this.fOrderService.destroy();
-            this.pOrderService.addToOrder(OrderItem);
-        }else{
+        if (Utils.isValid(json) && Utils.isJsonString(json)) {
+            this.confirmationDialogService.confirm('Please confirm', 'You have a active food cart. Adding this product to the cart will remove food order. Please confirm ?')
+                .then((confirmed) => {
+                    console.log('User confirmed:', confirmed);
+                    this.fOrderService.destroy();
+                    this.pOrderService.addToOrder(OrderItem);
+                })
+                .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+        } else {
             this.pOrderService.addToOrder(OrderItem);
         }
     }
 
-    addToFoodOrder(foodItem: FoodOrderItem){
+    addToFoodOrder(foodItem: FoodOrderItem) {
         console.log('Adding a food to cart');
         var json = this.localService.getData(Constants.StorageItem_P_Order);
-        if ( Utils.isValid(json) && Utils.isJsonString(json)){
-            this.toastService.warning("You have a active product cart. Please be informed that product cart will be removed.");
-            this.pOrderService.destroy();
-            this.fOrderService.addToOrder(foodItem);
-        }else{
+        if (Utils.isValid(json) && Utils.isJsonString(json)) {
+            this.confirmationDialogService.confirm('Please confirm', 'You have a active Product cart. Adding this food to the cart will remove Prouct order. Please confirm ?')
+                .then((confirmed) => {
+                    console.log('User confirmed:', confirmed);
+                    this.pOrderService.destroy();
+                    this.fOrderService.addToOrder(foodItem);
+                })
+                .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+        } else {
             this.fOrderService.addToOrder(foodItem);
         }
     }
